Move `sync_from_checkpoint` to new module
